requirements-dev.txt
-r requirements.txt

selenium==4.9
black==23.3
django_extensions==3.2
ipython==8.13.2
coverage==7.2.5
pytest==7.4.0
pytest-django==4.5.2
pre-commit==3.7.0
isort==5.13.2